+++ bedriftshelsetjenesten/src/BedriftshelsetjenesteGodkjenningsregister/Kontaktinformasjon.java
@@ -2,7 +2,6 @@
  */
 package BedriftshelsetjenesteGodkjenningsregister;
 
-import java.math.BigDecimal;
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.EObject;
 
@@ -32,12 +31,12 @@ public interface Kontaktinformasjon extends EObject {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Telefon</em>' attribute.
-	 * @see #setTelefon(BigDecimal)
+	 * @see #setTelefon(int)
 	 * @see BedriftshelsetjenesteGodkjenningsregister.BedriftshelsetjenestenPackage#getKontaktinformasjon_Telefon()
-	 * @model dataType="BedriftshelsetjenesteGodkjenningsregister.Telefon8Siffer"
+	 * @model
 	 * @generated
 	 */
-	BigDecimal getTelefon();
+	int getTelefon();
 
 	/**
 	 * Sets the value of the '{@link BedriftshelsetjenesteGodkjenningsregister.Kontaktinformasjon#getTelefon <em>Telefon</em>}' attribute.
@@ -47,7 +46,7 @@ public interface Kontaktinformasjon extends EObject {
 	 * @see #getTelefon()
 	 * @generated
 	 */
-	void setTelefon(BigDecimal value);
+	void setTelefon(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Epost Adresse</b></em>' attribute.
@@ -94,35 +93,35 @@ public interface Kontaktinformasjon extends EObject {
 	void setInternettAdresse(String value);
 
 	/**
-	 * Returns the value of the '<em><b>Postadresse</b></em>' reference.
+	 * Returns the value of the '<em><b>Postadresse</b></em>' containment reference.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Postadresse</em>' reference.
+	 * @return the value of the '<em>Postadresse</em>' containment reference.
 	 * @see #setPostadresse(Postadresse)
 	 * @see BedriftshelsetjenesteGodkjenningsregister.BedriftshelsetjenestenPackage#getKontaktinformasjon_Postadresse()
-	 * @model
+	 * @model containment="true"
 	 * @generated
 	 */
 	Postadresse getPostadresse();
 
 	/**
-	 * Sets the value of the '{@link BedriftshelsetjenesteGodkjenningsregister.Kontaktinformasjon#getPostadresse <em>Postadresse</em>}' reference.
+	 * Sets the value of the '{@link BedriftshelsetjenesteGodkjenningsregister.Kontaktinformasjon#getPostadresse <em>Postadresse</em>}' containment reference.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * @param value the new value of the '<em>Postadresse</em>' reference.
+	 * @param value the new value of the '<em>Postadresse</em>' containment reference.
 	 * @see #getPostadresse()
 	 * @generated
 	 */
 	void setPostadresse(Postadresse value);
 
 	/**
-	 * Returns the value of the '<em><b>Kontaktperson</b></em>' reference list.
+	 * Returns the value of the '<em><b>Kontaktperson</b></em>' containment reference list.
 	 * The list contents are of type {@link BedriftshelsetjenesteGodkjenningsregister.Kontaktperson}.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Kontaktperson</em>' reference list.
+	 * @return the value of the '<em>Kontaktperson</em>' containment reference list.
 	 * @see BedriftshelsetjenesteGodkjenningsregister.BedriftshelsetjenestenPackage#getKontaktinformasjon_Kontaktperson()
-	 * @model
+	 * @model containment="true"
 	 * @generated
 	 */
 	EList<Kontaktperson> getKontaktperson();
